refactor: move wecom under channel
