+++ CLAUDE.md
@@ -9,7 +9,7 @@ TinyClaw is now organized around the Core Model:
 - `internal/core`: Room, Message, Invocation, Delivery, and Trigger Policy.
 - `internal/storage`: PostgreSQL persistence for the Core Model.
 - `internal/api`: HTTP adapter for inbound messages, invocation completion/failure, and delivery polling/ack.
-- `wecom/`: legacy WeCom SDK helpers and clients.
+- `channel/wecom/`: legacy WeCom SDK helpers and clients.
 
 The old in-repo TypeScript agent runtime, sandbox orchestrator, and `RoomChat` gRPC bridge have been removed. Do not reintroduce sandbox or agent execution code unless a new sandbox design is explicitly accepted.
