+++ channel/wecom/finance/lib/tool_testSdk.cpp
