+++ channel/wecom/finance/lib/version.txt
