+++ channel/wecom/finance/sdk_linux.go
