+++ channel/wecom/finance/sdk_unsupport.go
