+++ channel/wecom/user.go
